requirements.txt
Flask
Flask-Cors
Flask-Session
face_recognition
numpy
opencv-python
google-generativeai
markdown
PyMuPDF
Pillow